Make gryzor hack working (ICMP reject):  * NuAuth store 28 first bytes of each packet (needed for ICMP reject)  * Set payload_len attribute to 16 bits (instead of 32)    for nuauth_decision_response_t structure  * Minimum packet size is 28 bytes for get_ip_headers()

+++ src/include/proto.h
@@ -266,9 +266,9 @@ typedef struct {
     uint8_t priority;           /*!< Priority ? */
     uint16_t padding;           /*!< Padding (0x0000) */
     uint32_t packet_id;         /*!< NetFilter packet unique identifier */
-    uint32_t payload_len;      /*!< Indicate the length of datas in the recv buffer after 
+    uint16_t payload_len;       /*!< Indicate the length of datas in the recv buffer after 
                                   the end of the structure that contains the payload of packet. Set
-                                 to 0 to treat the following datas as a new decision response */
+                                  to 0 to treat the following datas as a new decision response */
 } nuauth_decision_response_t;    
 
 #endif
